Added API (with many missing features), ready for some basic integration tests
